refactor: moved testKeyGenerator()
